Issue #7: Fix bugs #14 and #13.
Bug #14: An error in the 'decode' method was causing a problem in assigning a name to the state, which was hindering the step-by-step debugging process.

Bug #13: The transition matrix contained subsequent states that used the same event for transitioning. For instance, there was a transition from state 0 to 1 with event 32, and another transition from state 1 to 3 with the same event 32. This created the impression of an error, when in reality it was due to a rapid transition.

Change the class of properties from the classes StateMachine and StateMachineUtils from int32 to double to be easy to accept data.